refactor(language-core): decoupled from typescript
